Create display class to separate functionality
